Patch Commands.pm

+++ lib/setup-storage/Commands.pm
@@ -337,6 +337,8 @@ sub build_btrfs_commands {
     }
   }
 
+  my %mkfs_done;
+
   foreach my $config (keys %FAI::configs) { # loop through all configs
     next unless ($config eq "BTRFS");
 
@@ -365,16 +367,23 @@ sub build_btrfs_commands {
       $pre_req = "exist_" . $devs[0];
     }
     # creates the BTRFS volume/RAID
-    if ($raidlevel eq 'single') {
-          &FAI::push_command("mkfs.btrfs -d single $createopts ". join(" ",@devs),
-                             "$pre_req",
-                             "btrfs_built_raid_$id");
-
-        } else {
-          &FAI::push_command("mkfs.btrfs -d raid$raidlevel $createopts ". join(" ",@devs),
-                             "$pre_req",
-                             "btrfs_built_raid_$id");
-        }
+     if ($raidlevel eq 'single') {
+      if (exists $mkfs_done{join(" ", @devs)}) {
+        &FAI::push_command("true",
+          "$pre_req",
+          "btrfs_built_raid_$id");
+      } else {
+        print "Adding mkfs command for '", join(", ", @devs), "'.\n";
+        &FAI::push_command("mkfs.btrfs -d single $createopts ".join(" ",@devs),
+          "$pre_req",
+          "btrfs_built_raid_$id");
+          $mkfs_done{join(" ", @devs)} = '1';
+      }
+    } else {
+      &FAI::push_command("mkfs.btrfs -d raid$raidlevel $createopts ".join(" ",@devs),
+        "$pre_req",
+        "btrfs_built_raid_$id");
+    }
 
     # initial mount, required to create the initial subvolume
     &FAI::push_command("mount $devs[0] /mnt",
